requirements.txt
Flask==2.3.0
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.0
pydantic>=2.0.0
google-generativeai>=0.8.0
groq>=0.11.0
